d_a_npc_gk1: _create 98.96->99.37% (createInit returns u8 - caller clrlwi, a_size_tbl->a_siz_tbl name match)

+++ include/d/actor/d_a_npc_gk1.h
@@ -22,7 +22,7 @@ public:
     void _nodeCB_Neck(J3DNode*, J3DModel*);
     void _nodeCB_BackBone(J3DNode*, J3DModel*);
     s32 init_GK1_0();
-    s32 createInit();
+    u8 createInit();
     void play_animation();
     void setMtx(bool);
     int bckResID(int);
